requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-generativeai==0.3.2
python-dotenv==1.0.1
sqlalchemy==2.0.27
pytest==8.0.2
httpx==0.26.0
pytest-asyncio==0.23.5
pydantic==2.6.1 